- Fix service_flags compile warning. - Add support for chanfix service flags in operator blocks.
git-svn-id: http://svn.ratbox.org/svnroot/rserv/trunk@27029 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ include/service.h
@@ -14,15 +14,15 @@ struct cachefile;
 
 struct service_command
 {
-        const char *cmd;
-        int (*func)(struct client *, struct lconn *, const char **, int);
+	const char *cmd;
+	int (*func)(struct client *, struct lconn *, const char **, int);
 	int minparc;
 	struct cachefile **helpfile;
-        int help_penalty;
-        unsigned long cmd_use;
+	int help_penalty;
+	unsigned long cmd_use;
 	int userreg;
 	int operonly;
-	unsigned long long operflags;
+	uint64_t operflags;
 };
 
 struct service_handler
